Change `port` from uint16_t to uint32_t, to support VSOCK (#583)

+++ include/aws/crt/mqtt/Mqtt5Client.h
@@ -474,7 +474,7 @@ namespace Aws
                  *
                  * @return this option object
                  */
-                Mqtt5ClientOptions &WithPort(uint16_t port) noexcept;
+                Mqtt5ClientOptions &WithPort(uint32_t port) noexcept;
 
                 /**
                  * Set booststrap for mqtt5 client
@@ -754,7 +754,7 @@ namespace Aws
                 /**
                  * Network port of the MQTT server to connect to.
                  */
-                uint16_t m_port;
+                uint32_t m_port;
 
                 /**
                  * Client bootstrap to use.  In almost all cases, this can be left undefined.
